requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
boto3==1.35.71
botocore==1.35.71
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
envs==1.4
flexcache==0.3
flexparser==0.4
frozenlist==1.5.0
humanize==4.11.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
multidict==6.1.0
otf-api==0.8.1
Pint==0.24.4
platformdirs==4.3.6
propcache==0.2.0
pycognito==2024.5.1
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.0
